+++ src/AdvancedPS.jl
@@ -2,7 +2,6 @@ module AdvancedPS
 
 using AbstractMCMC: AbstractMCMC
 using Distributions: Distributions
-using Libtask: Libtask
 using Random: Random
 using StatsFuns: StatsFuns
 using Random123: Random123
