+++ src/resampling.jl
@@ -88,9 +88,9 @@ generated by stratified resampling.
 
 In stratified resampling `n` ordered random numbers `u₁`, ..., `uₙ` are generated, where
 
-``uₖ \\sim U[(k - 1) / n, k / n)``. 
+``uₖ \\sim U[(k - 1) / n, k / n)``.
 
-Based on these numbers the samples `x₁`, ..., `xₙ` are selected according to 
+Based on these numbers the samples `x₁`, ..., `xₙ` are selected according to
 the multinomial distribution defined by the normalized `weights`, i.e., `xᵢ = j` if and only if
 
 ``uᵢ \\in \\left[\\sum_{s=1}^{j-1} weights_{s}, \\sum_{s=1}^{j} weights_{s}\\right)``.
@@ -137,11 +137,11 @@ Return a vector of `n` samples `x₁`, ..., `xₙ` from the numbers 1, ..., `len
 generated by systematic resampling.
 
 In systematic resampling a random number ``u \\sim U[0, 1)`` is used to generate `n` ordered
-numbers `u₁`, ..., `uₙ` where 
+numbers `u₁`, ..., `uₙ` where
 
-``uₖ = (u + k − 1) / n``. 
+``uₖ = (u + k − 1) / n``.
 
-Based on these numbers the samples `x₁`, ..., `xₙ` are selected according to 
+Based on these numbers the samples `x₁`, ..., `xₙ` are selected according to
 the multinomial distribution defined by the normalized `weights`, i.e., `xᵢ = j` if and only if
 
 ``uᵢ \\in \\left[\\sum_{s=1}^{j-1} weights_{s}, \\sum_{s=1}^{j} weights_{s}\\right)``
